chore: Move all the types of create org into the types folder
