+++ components/shared/CreateOrgForm.tsx
@@ -3,7 +3,6 @@
 import Link from "next/link"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useFieldArray, useForm } from "react-hook-form"
-import { z } from "zod"
 
 import { Separator } from "@/components/ui/separator"
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
@@ -19,52 +18,7 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import React from "react"
-
-// the zod schema for the form requirements
-const createOrgFormSchema = z.object({
-  companyName: z
-    .string()
-    .min(2, {
-      message: "companyName must be at least 2 characters.",
-    })
-    .max(70, {
-      message: "companyName must not be longer than 30 characters.",
-    }),
-  companyDomain: z
-    .string()
-    .min(2, {
-      message: "companyDomain must be at least 2 characters.",
-    })
-    .max(70, {
-      message: "companyDomain must not be longer than 30 characters.",
-    }),
-  size: z.enum(["1-10", "11-50", "51-100", "101-500", "501-1000", "1000+"], {
-    required_error: "You need to select size of your mentors",
-  }),
-  focus: z.enum(["religion", "software", "engineering", "healthTech", "crypto"], {
-    required_error: "You need to select size of your mentors",
-  }),
-  role: z.enum(["ceo/owner", "hrManager", "it/techManager", "it/techStaff", "hrStaff", "other"], {
-    required_error: "You need to select size of your mentors",
-  }),
-  otherRole: z
-    .string()
-    .min(2, {
-      message: "specificRole must be at least 2 characters.",
-    })
-    .max(30, {
-      message: "specificRole must not be longer than 30 characters.",
-    })
-    .optional()
-})
-
-type createOrgFormValues = z.infer<typeof createOrgFormSchema>
-
-
-interface CreateOrgFormProps {
-  currentPage: string,   // The current active page (either "first" or "second")
-  mySubmit: (data: createOrgFormValues) => void
-}
+import { CreateOrgFormProps, createOrgFormSchema, createOrgFormValues } from "@/types/create-org"
 
 const sizeOptions = [
   { label: "1-10", value: "1-10" },
