+++ components/views/auth/create-org.tsx
@@ -3,9 +3,10 @@
 import CreateOrgForm from "@/components/shared/CreateOrgForm";
 import CreateOrgSidebar from "@/components/shared/CreateOrgSidebar"
 import LandingPageHeader from "@/components/shared/LandingPageHeader"
+import { Page } from "@/types/create-org";
 import { useState, useRef } from "react"
 
-export type Page = "first" | "second";
+
 
 const CreateOrgView = () => {
   const [currentPage, setCurrentPage] = useState<Page>("first");
@@ -27,7 +28,7 @@ const CreateOrgView = () => {
       <LandingPageHeader showButton={false} />
 
       <div className="flex-1 flex w-screen">
-        <CreateOrgSidebar triggerSubmit={triggerFormSubmit} currentPage={currentPage} handleSetCurrentPage={(page) => setCurrentPage(page)}/>
+        <CreateOrgSidebar triggerSubmit={triggerFormSubmit} currentPage={currentPage} handleSetCurrentPage={(page: Page) => setCurrentPage(page)}/>
 
         {/* where the organization form will be added */}
         <div className="flex-1 bg-gray-100 flex justify-center items-center p-10">
