+++ package.json
@@ -32,7 +32,7 @@
     "lucide-react": "^0.460.0",
     "next": "15.0.3",
     "next-auth": "^4.24.10",
-    "next-themes": "^0.4.3",
+    "next-themes": "^0.4.4",
     "openssl": "^2.0.0",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-dom": "19.0.0-rc-66855b96-20241106",
